Struct nft_benefits_vault::Contract

pub struct Contract { /* private fields */ }

Core structure of the smart contract.

Implementations

impl Contract

pub fn withdraw_all(
&mut self,
nft_contract_id: AccountId,
nft_id: TokenId
) -> Promise

Public function to withdraw all FTs from the Vault. The contract will check ownership of the NFT spectified by the arguments. After that it will try to find the vault with access via provided contract/id pair. Exactly 1 yoctoNEAR must be attached.

pub fn withdraw(
&mut self,
nft_contract_id: AccountId,
nft_id: TokenId,
fungible_token: AccountId
)

Public function to withdraw a single type of FTs from the Vault. The contract will check ownership of the NFT spectified by the arguments. After that it will try to find the vault with access via provided contract/id pair. And it makes ft_transfer with all available tokens on the provided fungible_token. Exactly 1 yoctoNEAR must be attached.

pub fn withdraw_amount(
&mut self,
nft_contract_id: AccountId,
nft_id: TokenId,
fungible_token: AccountId,
amount: U128
)

pub fn withdraw_all_callback(
&mut self,
nft_info: Result<Option<Token>, PromiseError>,
nft_id: NftId
)

Callback invokes after request to the NFT contract to check ownership and grant access to the vault. Private: can be called only by this contract.

Panics if:

called NOT by itself,

previous Promise failed,

there is no NFT with provided id on the NFT contract,

signer is NOT an NFT owner,

there is no Vault with access via given [NftId].

pub fn withdraw_callback(
&mut self,
nft_info: Result<Option<Token>, PromiseError>,
nft_id: NftId,
fungible_token: AccountId,
amount: Option<U128>,
replenish: bool
)

Callback makes transfer of a single FT type to the signer if all requirements are met.

pub fn add_replenishment_callback(
&mut self,
nft_contract_id: AccountId,
nft_id: TokenId,
callback: String,
args: String
)

This method is used to add replenishers for the Vault. When NFT owner calls withdraw methods the Contract will transfer available tokens from its own balance and will request available benefits from registered replenishers. This method MUST be called and signed by replenisher himself because future replenishment requests will be made to the signer account.

pub fn adjust_balance(
&mut self,
res: Result<(), PromiseError>,
nft_id: NftId,
fungible_token: AccountId,
amount: U128
)

Callback invokes after each FT transfer call from this contract in withdrawal process. If transfer was success, internal balance will be reduced by the amount transferred. Private: can be called only by this contract.

impl Contract

pub fn balance_of(
&self,
nft_contract_id: AccountId,
nft_id: TokenId
) -> Option<BalanceView>

Function to view the full balance of the vault.

pub fn vault(
&self,
nft_contract_id: AccountId,
nft_id: TokenId
) -> Option<VaultView>

Function to view the amount of assets in the Vault.

pub fn replenishers(
&self,
nft_contract_id: AccountId,
nft_id: TokenId
) -> Option<Vec<Replenisher>>

Function to view the replenishers of the vault.

impl Contract

pub fn ext(account_id: AccountId) -> ContractExt

API for calling this contract’s functions in a subsequent execution.

impl Contract

pub fn new() -> Self

Trivial init function. Panics if the contract is already initialized.

impl Contract

pub fn store(&mut self, nft_id: &NftId, fungible_token: &AccountId, amount: u128)

Adds provided amount of tokens to the vault specified by [NftId]. If there is no vault for the provided [NftId] then it will create a new one.

pub fn get_vault(&self, nft_id: &NftId) -> Vault

Shortcut to get vault from internal storage. Panics if there is no vault associated with the given [NftId].

pub fn get_vault_or_create(&self, nft_id: &NftId) -> Vault

Returns existing or new vault.

pub fn transfer_to(
fungible_token: AccountId,
nft_owner: AccountId,
amount: u128
) -> Promise