Struct nft_benefits_vault::ContractExt

pub struct ContractExt { /* private fields */ }

Implementations

impl ContractExt

pub fn ft_on_transfer(
self,
sender_id: AccountId,
amount: U128,
msg: String
) -> Promise

impl ContractExt

pub fn balance_of(self, nft_contract_id: AccountId, nft_id: TokenId) -> Promise

Function to view the full balance of the vault.

pub fn vault(self, nft_contract_id: AccountId, nft_id: TokenId) -> Promise

Function to view the amount of assets in the Vault.

pub fn replenishers(self, nft_contract_id: AccountId, nft_id: TokenId) -> Promise

Function to view the replenishers of the vault.

impl ContractExt

pub fn with_attached_deposit(self, amount: Balance) -> Self

pub fn with_static_gas(self, static_gas: Gas) -> Self

pub fn with_unused_gas_weight(self, gas_weight: u64) -> Self

impl ContractExt

pub fn new(self) -> Promise

Trivial init function. Panics if the contract is already initialized.

pub fn withdraw_all(self, nft_contract_id: AccountId, nft_id: TokenId) -> Promise

Public function to withdraw all FTs from the Vault. The contract will check ownership of the NFT spectified by the arguments. After that it will try to find the vault with access via provided contract/id pair. Exactly 1 yoctoNEAR must be attached.

pub fn withdraw(
self,
nft_contract_id: AccountId,
nft_id: TokenId,
ft_contract_id: AccountId
) -> Promise

Public function to withdraw a single type of FTs from the Vault. The contract will check ownership of the NFT spectified by the arguments. After that it will try to find the vault with access via provided contract/id pair. And it makes ft_transfer with all available tokens on the provided ft_contract_id. Exactly 1 yoctoNEAR must be attached.

pub fn add_replenishment_callback(
self,
nft_contract_id: AccountId,
nft_id: TokenId,
callback: String,
args: String
) -> Promise

This method is used to add replenishers for the Vault. When NFT owner calls withdraw methods the Contract will transfer available tokens from its own balance and will request available benefits from registered replenishers. This method MUST be called and signed by replenisher himself because future replenishment requests will be made to the signer account.

pub fn withdraw_all_callback(self, nft_id: NftId) -> Promise

Callback invokes after request to the NFT contract to check ownership and grant access to the vault. Private: can be called only by this contract.

Panics if:

called NOT by itself,

previous Promise failed,

there is no NFT with provided id on the NFT contract,

signer is NOT an NFT owner,

there is no Vault with access via given [NftId].

pub fn withdraw_callback(
self,
nft_id: NftId,
ft_contract_id: AccountId
) -> Promise

Callback makes transfer of a single FT type to the signer if all requirements are met.

pub fn adjust_balance(
self,
nft_id: NftId,
ft_contract_id: AccountId,
amount: u128
) -> Promise

Callback invokes after each FT transfer call from this contract in withdrawal process. If transfer was success, internal balance will be reduced by the amount transferred. Private: can be called only by this contract.